gcp-serverless-function/requirements.txt
Flask
pandas
joblib
google-cloud-storage
pycaret